pyproject.toml
[tool.poetry]
name = "past-mtl-monitors"
version = "0.1.0"
description = "A library for creating past metric temporal logic monitors."
repository = "https://github.com/mvcisback/past-mtl-monitors"
readme="README.md"
authors = ["Marcell Vazquez-Chanlatte <mvc@linux.com>"]

[tool.poetry.dependencies]
python = "^3.8"
attrs = "^19.3.0"
discrete-signals = "^0.7.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
hypothesis = "^5.5.1"
codecov = "^2.0.15"
pytest-flake8 = "^1.0.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
